Update how prize/tier chances work
The max range by default is now 100 and is removed from the crate files. You can now use decimals like 0.5 or 0.005

+++ paper/src/main/java/com/badbones69/crazycrates/api/PrizeManager.java
@@ -178,9 +178,9 @@ public class PrizeManager {
         if (crate.getTiers() != null && !crate.getTiers().isEmpty()) {
             for (int stopLoop = 0; stopLoop <= 100; stopLoop++) {
                 for (Tier tier : crate.getTiers()) {
-                    int chance = tier.getChance();
+                    double chance = tier.getChance();
 
-                    int num = MiscUtils.useOtherRandom() ? ThreadLocalRandom.current().nextInt(tier.getMaxRange()) : new Random().nextInt(tier.getMaxRange());
+                    double num = MiscUtils.useOtherRandom() ? ThreadLocalRandom.current().nextDouble(tier.getMaxRange()) : new Random().nextDouble(tier.getMaxRange());
 
                     if (num >= 1 && num <= chance) {
                         return tier;
